Section::Resolve() aborts on error
Aborts immediately on duplicate providers.
Aborts on missing deps only if --strict is set.

We've been allowing missing deps even when --strict is on. (only
emitting errors) Now, linkerconfig aborts in strict mode when there's a
missing dep.

Because --strict has been used only in tests, not in production, this
won't break. However, it will be used by `init` on rebootless update of
vendor apex. Aborting on missing deps will reject installation of an
apex when the required libs are not available.

Bug: 232173613
Test: atest (in linkerconfig dir)
Change-Id: Ife00864a759b52b84498fb1b523b8a366a52801e