Vendor APEX do not link to VNDK when it is deprecated
From prior work, VNDK namespace was still active from Vendor APEXes
because some of them relied on VNDK libraries by linking. However, this
is not available when VNDK is deprecated. This change removes link from
Vendor APEXes to the VNDK when the VNDK is deprecated.

Bug: 290318998
Test: Checked from AOSP CF x86_64 Phone build with KEEP_VNDK=false that
VNDK namespace has been removed with vendor APEXes

Change-Id: I3bcb9aee6a16e25cdaa428572f34d15c18bd7f14