Check if device is treblelized only with ro.treble.enabled property
Currently there are some codes in the linkerconfig which assumes that
device is treblelized iff ro.vndk.version is set. However, this will be
no longer true once VNDK is deprecated and ro.vndk.version is set as
empty. This change updates linkerconfig to check if device is
treblelized only with ro.treble.enabled property.

Bug: 290159430
Test: There was no change with rundiff.sh
Change-Id: Ie49f9f642909581e2d24507621e521c0277588b7