Product is treblelized if the device is treblelized
There are some parts in linkerconfig which checks if product partition
is treblelized with ro.product.vndk.version. However, this is no longer
true, because once VNDK is deprecated then ro.product.vndk.version will
be empty even when product partition is treblelized. For now, we can
assume that all treblelized devices are product treblelized. This CL
changes the logic to check if product partition is treblelized by
checking if the device is treblelized.

Bug: 290159430
Test: rundiff.sh --update confirmed that there is no misgenerated
ld.config.txt file

Change-Id: I163ad1f9dbb539120e9a429724f0e785e71be79c